More flexible routing system
Introduced Route.method to bind a callback to an arbitrary method and a
rule. It is now possible to bind callback to non-standard HTTP method.

Added Route.regex which will bind a callback to a custom method and
regular expression. It can be nicely combined with Regex literal in
latest version of Vala.

Refactored all calls upon Route.route which will bind a Route instance
to an arbitrary HTTP method.