Uses GLib types over Gee in the framework.
We should minimize the dependency of the framework and opt for
high-performance data structure.

Using GLib.Queue for routes in Router as we need fast append and fast
traversal.

Checks if the routes contains the method in handler before accessing the
routes for a method. This would have led to a NULL exception if a method
wouldn't have been declared at all and accessed by user.

Gee should be reserved for examples and CTPL integration.