Uses a string? to represent the session content.
It is not right to assume that the session content is encoded according to
x-www-urlencoded as it could be pretty much anything a user would like to store.
It must at least be decoded from the environment variable to its string
representation.

This is also a good benefit in term of performance as the session will be
decoded on need rather that systematically.

Encoding and decoding is now the responsibility of the application.

Updates the documentation consequently.