Experimental icon!
I am proposing this icon as we do not have any yet! It is a simple and red
oblique V using the DejaVu Sans font. The source is in svg and a 128x128 png is
served in in the default application.

The default app route for serving file have been slightly modified to serve
non-minified files such as icons.