Improve error handling for FastCGI implementation.
Adds error code in the message and set the exit status based on what the
application handler throws.

Error thrown from the application are caught. The code is used for the exit
status and message is dumped in the error stream.

Error handling for failed read operations.

OutputStream.flush is virtual, so it is a better approach to override it and
make it flush the FastCGI stream under the hood.

The error in write_headers is handled in write rather that having duplicated
error handling.