Uses a delegator to define VSGI.Application.
Application can be summarized as a callback handling a Request and Response
couple, so it is much more flexible to define it as a delegator.

This way, even Valum is not necessary to write basic web application:

    new SoupServer ((req, res) => {
        res.write ("Hello world!".data);
    }).run (args);

This way, any VSGI Application can be assigned as a Route handler, opening
interesting composition:

    app.get ("", app.handler); // recursive definition :)

    app.get ("", other_app.handler);

    app.get ("/scoped/here/<any:path>", other_app.handler);

It could be possible, for instance, to compose two frameworks written upon VSGI
specification.