Route.from_regex supports scoping.
Regular expression must however start matching after the scopes and leading '/'
so that we can prepend the scope and slash on the provien pattern.

Regex are also anchored since they must always do a full match of the input
URI, so it does not have to be explicitly given. This allow us to assume that
the regex does not start with a '^' and the '$' is appended for consistency.

Tests were slightly adapted:

 - no need to start a regex with a slash '/'
 - no need for anchors ('^' and '$')
 - new test for a scoped regular expression

Updates the documentation and examples to follow the changes.