Improve the documentation consistency.
More emphasis on the populate if match principle for matchers.

Uniform and more consistent use of heading (# rather than = and -) since we
often have more than 2 levels of heading.

mkdocs does not support Vala embedded code, so it is better to use JavaScript
for the moment.

Documents how futile it is to close streams with reference counting for typical
scenario. Also present some scenario where it is highly important like heavy
processing when the stream processing is already done.