VSGI.Test implementation following transparent namespace pattern.
Refactors the testing VSGI implementation under the VSGI.Test namespace.
Contairly to common implementations, Request and Response are public since they
are used as studs.

Removes the test_soup.vala and test_fastcgi.vala as they were unused and should
not be part of Valum testsuite, but more in a future VSGI testsuite.