Better integration of GLib.Application
Uses handle_local_option and activate instead of the command_line signal to
launch the server processing. This is more appropriate since a server is an
active instance and does not work sequentially like a CLI utility.

Uses hold and release from GLib.Application to keep a count of processing
requests. The server can now timeout due to its inactivity.

VSGI.Application handle function is an asynchronous function to provide
fundamentally asynchronous processings on user requests.

Router define setup and teardown, functions called before and after the
processing of a request. This is to replace connect and connect_after on the
previous handle signal.

setup default handler will fill the response with decent defaults such as status
code, content type and cookies.