Uses real stream by VSGI implementations.
It is necessary to use a property (or attribute) to represent the stream as
wrapping is not sufficient to provide full-access to the stream.

 - the Request InputStream is accessed through a body property
 - the Response OutputStream is accessed through a body property

Soup implementation uses 'steal_connection' to obtain access to the IOStream
underlying a Message. This way, it is possible to work with real GIO streams,
not just wrappers.

The connection InputStream is already consumed, so a simple MemoryInputStream is
used over the MessageBody data.

ChunkedOutputStream is introduced to provide chunked transfer encoding for Soup
implementation. The MessageBody was usually handling that part, but since we use
the stream directly, it is necessary to reimplement that part.

FastCGI implementation will use a wrapper as there is no way (yet to be found)
to access the stream.

Writing status line and headers has to be controlled by the response, thus these
functions are introduced and invoked on the first body access:

 - write_status_line to write status line
 - write_headers to write response headers

In order to 'close' the response when its processing is done before it gets
finalized, the 'end' signal is introduced. This way, when processing is going
down asynchronous calls, it will eventually call end on the response and notify
properly that the processing is done.

 - VSGI implementations will connect a callback to the end signal in order to
   print a message and free some related resources
 - at any point in the processing, a callback can be connected to perform
   operation at this important event

Remove some operations that were not really asynchronous

 - handle_async is useless since we want to let the user 'go asynchronous'
 - splice_async does nothing really interesting and it will probably never have
   asynchronous capabilities

test