Renames raw_body of base_stream in Response.
It doesn't hold a body but more a general stream containing status line, headers
and a body.

Adds the possibility to filter or redirect the Response body by setting the
property with a new value.

Adds base_stream in Request and provide the possibility to filter or redirect
the body the same way it works with the Request.

Fixes tests by using stubbed MemoryInputStream and MemoryOutputStream for the
base_stream property.