Introduces Next in Router.
Route.Handler has now a third argument which is a callback used to keep routing.

The setup signal does not serve any purpose, so it has been removed.

the teardown signal will eventually be removed in the 0.2.* release since we
will have the possibility to connect a callback to do the teardown processings.

Removes the timer examples as it requires the setup signal.

Removes the setup and teardown documentation section as it will be rewritten
properly to cover how to do that with the next() function.