Properly close the request and response body in end default handler.
The base_stream behind Request and Response must be closed from their most
enclosing stream, which is the body property.

Closing the base stream is dangerous as some converters might not be notified
and omit to write important closing data. This is the case for the
ChunkedConverter, as it must write an empty chunk to end the chunking.

It is still possible to close streams asynchronously within the application
handler, close will not be effective if the stream is already closed.

Documents that behiaviour and adds an example of asynchronously closed body.