Ends in a continuation instead of a signal in Response.
Moves the Response end signal in a continuation passed in Application handle
function. The function has to be invoked by the application in a synchronous or
asynchronous context.

Adapts the routing to pass the end continuation over process_routing.

Updates examples to take the third argument end and call it instead of res.end.

Updates tests to take that change into consideration when invoking
Router.handle.