Provide error code and message to the status handler.
Keys 'code' and 'message' are setted in the request parameters so that the
status handler can access that information.

The request parameters are initialized if still null, which might happen if no
matcher has populated them.

Optimally, that information would go through a state, but this is not really
easy to do without making major modifications to the router.