WIP: Uses reference counting to free connection
Replaces the base_stream property which was used to hold the raw stream by a
IOStream representing a connection.

It has considerable advantages:

 - the connection can wrap additionnal resources for the request and free them
   properly
 - it is compatible with GLib.SocketConnection
 - resources are reference-counted and freed as soon as possible
 - no need to explicitly end the processing

For libsoup-2.4 (<2.50), an implementation of IOStream will wrap the request and
response messages with memory streams and perform pause/resume on the message
automatically.

The 'end' signal does not have any utility as it can not be implemented with a
reference counting approach. When the connection is out of scope, request
resources are freed and it happens whenever the processing ends.

There is no way to produce a debug message when a request has been processed,
but this information can be obtained from the used HTTP technology.