Supports informational and success status codes.
It is useful to throw informational or success status codes and perform
generic handling for them. For instance, the CREATED status can be
thrown with the 'Location' header to indicate the newly available
resource.

Updates the HandlerCallback and NextCallback delegates to throw
Informational and Success error domains.

Updates process_routing to throw the new error domains.

Adds status codes that are not described in Soup.Status enumeration but
referenced in RFC documents and provide basic handling for some of them
that requires headers to be setted.

Updates the documentation to take these changes into consideration.

Adds tests to cover the introduced status codes and ensure that
appropriate headers are set by the router default handler.