NextCallback propagates a state to the next handler.
The invocation of the next continuation can be made with a state to
cover cases where a route would process data and another would transmit
it in the desired format.

The router will propagate the state in case a middleware wouldn't, so
that it does not become an obligation to pass the state in the next
call.

The state is represented by an GLib.Object and available as an argument
of HandlerCallback. It must be casted in the desired type, but it can be
assumed that the route knows of what type the state is as it is part of
a typical routing.

Adds a test to ensure that the state is propagated and updates the
documentation to highlight these changes.