Removes cookies from VSGI.
Cookies are stored in HTTP headers, so it's not really relevant for VSGI
to provide helpers for them.

Instead, the code to extract cookies is moved in
'Cookies.from_request_headers' function to provide equivalent parsing
capabilities. The functions provided by libsoup-2.4 for that purpose
takes a Soup.Message as parameter, which is not common to all VSGI
implementations.

Removes the code transfeering the cookies in the Router. The cookies are
set once with 'Set-Cookie' and sent back by the client, so they do not
have to be setted for every requests.

Adds an example to interact in a key-value way with cookies.

Adds a few tests to ensure that 'Cookies.from_request_headers' works
properly.

Migrates the documentation section about cookies in a document part of
Valum documentation.