Explicitly closes with async operations.
When performing asynchronous operations involving the response body, the
connection is freed before the body stream, which may result in a corrupted
response.

The typical example is ocurring with the ChunkedConverter which write a
zero-sized chunk just before the stream get closed. If the inner stream is
closed before, the client will never receive it.

The approach proposed is to close manually (or using the CLOSE_TARGET flag when
splicing) to ensure the correct order.

Updates documentation and examples to reflect these changes.