Exposes head_written as a property in Response.
It provides an additionnal information to know if the status line and
headers have been written and the possibility to connect a callback when
the property changes. The setter remains protected so that it can only
be assigned by the implementation.

Adds pre/post conditions to ensure that write_head and write_head_async
behaves correctly.

Adds a few tests for the newly exposed property.