Fixes libsoup-2.4 (<2.50) output stream operations.
The code was taking ownership of the message_body data, so the latter
couldn't send anything useful to the client. Instead, a simple
OutputStream interface is provided over a MessageBody like it was
before.

Response build_head function can now return null which result in writing
nothing in the response output stream. This is useful for implementation
that handles headers themselves.