Provides invoke in the Router to execute an arbitrary NextCallback
The function takes a request, response, NextCallback and an initial
state and invokes the callback in the routing context so that things
like status codes are handled properly.

This solve the kind of situation where next has to be invoked in an
async callback, but the exception handling is not available.

Router.handle uses invoke as en entry point so that the code for status
handling is reused.