Examples can be optionally built to avoid undesired dependencies.
Adds an option named '--enable-examples' to build examples instead of
forcing the user to install unnecessary dependencies. The flag is
documented in the installation document.

Docker and Travis CI does not build examples, so their dependencies has
been removed.

Docker uses ubuntu-debootstrap:latest to build in the smallest
environment possible.

No need to run or expose 3003 in the Dockerfile as it should only be
used to build a container with Valum pre-installed.