Documentation update and improvments.
Adds a section about the routing stack in the router document.

Rewords subrouting section to be more up-to-date with the current code.

 - no more setup or teardown (solved with next)
 - no more request parameters (deprecated in favour of routing stack)
 - uses 'null' instead of  '<any:any>'

Adds contextual note and warning to put emphasis on important
paragraphs.

Updates JSON code example to use the routing stack properly.

Replaces references to state in API docs to the routing stack.