Move cookies utilities back to VSGI.
Considering that libsoup-2.4 provides cookies utilities for its
'Soup.Message' class, it's rational that VSGI provides equivalent
functions to operate on 'Request' and 'Response'.

The utilities have been updated to work on 'Request' and 'Response'
instances rather that 'Soup.MessageHeaders' as cookies extraction
require a context to be performed correctly.

 - 'Cookies.from_request' extract cookies from the 'Cookie' headers
 - 'Cookies.from_response' extract cookies from the 'Set-Cookie' headers
   with the request URI as origin

The lookup utility operate on 'SList<Cookie>' to be sufficiently general
for request and response cookies.

Update the tests and documentation to cover the changes.