Router.methods and Router.all return created Route objects
This can be used to create routing sequence with 'then' in 'Route',
which was not possible before as the created Route objects remained
hidden in the Router.

    var routes = app.methods ({Request.GET}, "", (req, res) => {});

    // build a routing sequence for the 'GET' method
    routes[0].then ((req, res) => {

    });

Route.method is non-null as the change creates one Route object per
method, so we are not in the need for a 'null' state anymore.

Add tests to cover the returned Route objects from 'methods' and 'all'
in Router.