Request and Response filtering
Filters can perform various processing such as buffering, stream
redirection, immuability and much more...

Introduce 'RequestFilter' and 'ResponseFilter' which are abstract
classes to build filters similarly to 'GLib.FilterInputStream'. The
default implementation is to proxy the base object so that actual
filters can override desired properties.

Provide 'RequestConverter' and 'ResponseConverter' to apply converters
directly on request or response objects.

Modify 'NextCallback' to take a possibly filtered 'Request' and
'Response as input. When invoking next, the current request and response
must be passed if they are not filtered.

Adapt the examples to pass the request and response with 'next' and
update gzip example to use a converter.

Add a document about filters to cover filters and converters.