Documentation update for the final release
Reword request and response documents for better readability.

Remove references to Valum in VSGI and update code examples to use the
'Server' directly instead of the 'Router'.

Update usages of 'next' to pass the request and response objects.

Move scripting and persistence in recipes.

Use filters for the handling middleware documentation example.