Provide 'flatten' and 'flatten_async' in  Request
These two functions provide a handy shortcut to flatten the request body
stream into a 'uint8[]' buffer and set the buffer size appropriately.

If the encoding is set to CONTENT_LENGTH, a fixed-size buffer is used to
avoid reallocations.

The request body is automatically closed once the splice operation
finishes using the 'OutputStreamSpliceFlags.CLOSE_TARGET'.

The libsoup-2.4 implementation already accumulates the request body, so
we simply override the default behaviour.

Update examples that were splicing the request body with the new
functions.

Add a section in the request documentation.