Remove hard-dependency on GThread
SCGI use a 'SocketService' instead of a 'ThreadedSocketService' as VSGI
promotes the use of non-blocking I/O. This way, developer does not have
to deal with concurrency (synchronization, locking, ...) unless he
explicitly make use of threading.

Remove the '--max-threads' option from SCGI.

Remove the relevant warnings in the user documentation.