Generate user documentation in the 'build' folder
The build is triggered only if 'sphinx-build' is available.

Let Sphinx already handle already built sources by specifying
'always=True' in the task generator.

Remove 'docs/_build' from '.gitignore' as it will be part of the 'build'
directory, which is already ignored.

Remove the Travis CI script for building the user documentation as it is
already part of the build.