Fix cookie verification process
Use a time-constant comparison to avoid time-based attacks when
verifying a cookie signature.

Fix the @value argument assigned to the cookie value even if the cookie
is not verified. It has to keep a 'null' value as documented.

Add a test to cover bad cookie signature.

Update the documentation to cover the time-constant comparison and
reorder some blocks for better readability.