Introduce basic matching middlewares (not and noop)
The 'not' matching middleware provides a way to negate a passed matching
middleware to build more logical expressions:

    app.get ("")
       .and (not (cached))
       .then ((req, res) => { ... });

The handling callback for 'Valum.Router' helper methods is optional and
defaults to 'noop'. It is handy when the processing actually happens in
'then':

    app.get ("")
       .and (accept ("application/json"))
       .then ((req, res) => { ... });

Introduce the noop middleware that simply forward the request and
response to the next handler. It is used as a default value and
optimized away whenever possible.