Use a struct for 'Route'
Take advantage of a simpler definition for 'Route' by using a simple
struct that consist of a method, a matching callback and a handling
callbacks.

It reduces the memory usage considering that 'Route' instances are not
backed by GObject anymore.

The 'Route' object is no longer bound to a specific 'Router', which move
away the contextualization issues (like scoping) to how the matcher is
initialized.

It will also cover the need for inheritence as new kinds of 'Route' can
be implemented by specific static initializers.

The chaining is perfomed in 'Router' using 'Router.Builder': the created
'Route' object is available from the 'route' property.

The changes are only breaking 'Route' APIs, which are not typically
manipulated directly.

It is internal, so the user documentation does not have to be updated.

Update tests to take the change into consideration.