Introduce 'Router.rule' and move creation of 'Route' in 'Router'
Deprecate 'Router.method' in favour of 'Router.rule' to be more
consistent with 'Router.regex' and creation method in general.

Move route creation in 'Router' such that 'Router.rule' bind a
rule-based 'Route'and 'Router.regex' bind a regex-based 'Route'. There
is no more 'Route.from_rule' and 'Route.from_regex'.

Route creation does not need to capture the routing context since it
happens directly in the 'Router'.

Update tests to the new changes.