Fix #140 buffer and produce an EOF for the SCGI request body
Introduce 'SCGIInputStream' which filter a SCGI-compliant input stream
so that it produces an EOF when its content has been fully consumed.

The request body has to be fully consumed prior to produce the response,
so it is accumulated in the 'DataInputStream' buffer.

Rewrite the environment extraction code in a much cleaner way.

Add an example of splicing the request body into the response.