SCGI: Prefill the buffer with the netstring
The netstring is approximately 460B, so 512B of the stream is buffered
to perform the parsing in-memory.

Interpret the 'CONTENT_LENGTH' as a 'int64' to remain consistent with
how 'libsoup-2.4' deal with that quantity.

There is a possibility of truncating the request body on system where
sizeof (size_t) <  sizeof (int64) if its length cannot be stored in a
'size_t'.

Call 'fill' with '-1', which correspond to fill the internal stream
buffer.