Rename 'context' for 'stream_events'
The more general idea is that middlewares should be verbs in the
imperative tense, pretty much like git commits.

Unlike Express.js, middleware are not mounted on the router, but rather
part of the process, so it makes much more sense to name them as
actions.

    app.get ("", compress (ZlibCompressFormat.GZIP))
        .then (send_events ((req, send) => {
            send (null, "some event");
        }));

Update the code example in the documentation.