CGI: Handle the URI scheme and move environment processing
Move the code that checks the 'HTTPS' and 'REQUEST_URI' variables into
'CGI.Request'. It's a desirable behaviour to have conformity with
non-standard variables sent in real context.

It is sufficient to check if 'HTTPS' is non-empty to assume that the
request is sent securely.

Check the scheme in 'PATH_TRANSLATED' to detect usage of HTTPS.

Add the GIR namespace and version declaration for 'VSGI.CGI'.