Provide a middleware and utilities to handle subdomains
The 'subdomain' middleware matches request which subdomain is consistent
with the provided pattern. The pattern also support jokers '*'.

    var app = new Router ();
    var api = new Router ();

    app.matcher (subdomain ("api"), api.handle);

The default behaviour is to accept request which are part of the
provided subdomain in a suffix fashion. The 'strict' flag expects the
request to have the exact same subdomains.

The 'extract' utility is capable of splitting a hostname in an array of
labels. It drops the two top-most labels (eg. 'example.com') as they are
not typically part of what an application should consider as subdomains.

Provide minimal documentation and tests.