Partially revert '72792bb2170804c0ffd726746753c102e1868881'
Revert 'noop' and 'Route.and' in favour of the much cleaner way to
define matching process using logical clauses.

The clause is already functionally complete and does not require the
'Route.and' helper. With futur matching process, we could eventually
perform match such as:

    app.matcher (and (method ("GET"),
                      rule ("/"),
                      accept ("application/json")),
                 (req, res) => {
        res.body.write_all ("Hello world!".data, null);
    });

The 'Route.then' function is essential to build handling pipeline.