Build separate libraries for VSGI implementations
Organize sources in multiple folders to simplify the build and reduce
dependencies.

 - isolate the use of 'fcgi.h' for 'vsgi-fastcgi'
 - isolate the use of 'gio-unix-2.0' for 'vsgi-cgi'
 - provide a pkg-config file for each implementation

Each VSGI implementation is built as a 'TypeModule' and define the
'plugin_init' symbol to load its 'Server' implementation. This will be
used along with the loader described in #130.

Rename 'Soup' for 'HTTP' to avoid a *dirty* namespace conflict.

Extract the test implementation of VSGI under the name of 'VSGI.Mock'.