Decouple 'Route' with inheritence
Provide the follwing 'Route' implementations:

 - 'AnyRoute'
 - 'RegexRoute'
 - 'RuleRoute'
 - 'WildcardRoute' to match '*' path

Hide 'Router.types', it is injected in 'RuleRoute'. A more complex
structure might be required for typed parameters.

Fix #139 with 'WildcardRoute', the '*' rule is handled appropriately
when the scope stack is empty.

With isolation and inheritence, it will be easier to implement feature
like 'get_path' #45.

Remove 'Router' from 'Route' and reimplement 'then' without touching the
'Router', which avoid creating any 'Route' objects.