Move scoping, rooting and 'null' rule handling in 'Router'
'RuleRoute' and 'RegexRoute' take explicit path as input.

Automatically anchor the 'RegexRoute' with 'RegexMatchFlags.ANCHORED'
instead of prefixing and suffixing '^' and '$'.

Redefine the 'any' type to match empty string as well so that it can be
used for the 'null' rule.

Move tests to check the feature now provided by 'Router'.