Fix 'Route' and expose 'Router.route'
It's a good thing to let end-user define their own 'Route' objects.

Just set the 'RegexCompileFlags.OPTIMIZE' for 'Router.regex', it breaks
on the CI.

Hide the handler callback field in 'Route' and provide
'set_handler_callback' to assign it a new value.

Assign the 'rule' property in 'RuleRoute' and add some assertions to
ensure that it's in a correct state.

'RegexRoute' does not have to throw any 'RegexError' since it's simply
taking the provided regular expression as is.

Initialize the captures in 'RegexRoute' in a 'construct' block so that
subclasses can use the GObject constructor.

Update tests and code documentation.