FastCGI: Fix polling by subclassing 'UnixOutputStream'
Remove the 'GLib.Socket' and old polling code to let 'UnixOutputStream'
handle all that business.

Still, 'read' and 'write' are defined to read and write in the
multiplexed FastCGI stream.

Rewrite the initialization since we don't need the socket anymore, but
only a grab on the file descriptor.

Also, the file descriptor should not be member of 'Server' as we might
add support for multiple sources (see #132).