Rename 'Cookies' for 'CookieUtils' and improve 'sign'
VSGI is more about to provide utilities around 'Soup.Cookie', so the new
name is much more coherent.

Add a correct namespace annotation for GIR using CamelCase.

Change the 'sign' function to alter the provided cookie in-place: the
old behaviour was wasteful since you had to create a whole new cookie
with the signature.

Update tests and documentation consequently.