vsgi: Write the response head on disposal
The router cannot initialize with 'chunked' encoding, because it would
break empty responses. It must therefore be an explicit choice from the
end-user knowingly that something will be produced.

There's no need for the post-condition since we can be assured that the
head will be written on disposal.

Update the documentation to to include that behaviour.