vsgi: Provide expansion utilities to write the response body
Provide 'expand', 'expand_utf8' and their respective asynchronous
versions. It's similar to 'flatten' for the request body.

In short, it perform the following operations:

 - define the 'Content-Length' header
 - write the response head
 - expand the buffer into the response body

For fixed size response, it's preferable to 'body.write_all' because it
actually set the 'Content-Length'.

Update examples and documentation to use the utility, in particular
'expand_utf8'.