vsgi: Consistent options and messages for implementations
Capitalize option descriptions.

Use the 'fcgi' scheme for UNIX domain sockets and TCP port.

SCGI listen on both IPv4 and IPv6 interfaces, so we log both listened
addresses.

Use the 'scgi' scheme for TCP port and add a '-f' option to specify the
file descriptor on which to listen with SCGI.

Show the backlog for SCGI.

Update the SCGI example so that it can be passed CLI arguments.