vsgi: Convert 'Request' and 'Response' in-place
Filters and converters yield way too much complexity where all we want,
in the end, is a way to process the request and response body.

All that process is much better handled with 'GLib.Converter', so
instead we provide 'convert' methods in both 'Request' and 'Response'
to lazily apply the provided converter on the body.

Update tests and examples.