Replace status handlers with 'status' middleware
The 'status' middleware allow one to handle status raised by the
following handler.

    app.use (status (Soup.Status.NOT_FOUND, (req, res, next, ctx, err) => {
        // handle the status here...
    }));

    app.get ("/", () => {
        throw new ClientError.NOT_FOUND ("The requested URI '/' was not found.");
    });

Remove 'Router.status' in favour of a middleware based approach.

Do not treat other errors (eg. 'IOError') like a '500' status. Instead,
they have to be handled explicitly with a 'try' statement around an
invocation to the 'next' continuation.

The 'Context' is remaining untouched; the error is now passed as a
parameter to the middleware. There is no 'message' key anymore and the
actual original error can be inspected.

Status are forwarded upstream if the 'next' callback is invoked. It's
therefore possible to build a stack of status handlers.

Add a new 'status-test' test binary and a move the documentation into
'docs/middlewares/status.rst'.