Remove 'invoke' in favour of a 'basic' middleware
The middleware perform essentially the same work, but in a reusable way.
Also, status handling is not imposed anymore on the end-user since it
has to be explicitly mounted with 'use'.

Modify 'Router.perform_routing' to take a continuation to follow when
the recurrence finish. 'Router.handle' use that to generate a
'404 Not Found' or any appropriate error.

This change make it possible to handle the '404 Not Found' error
generated by the router using the 'status' middleware and alike.

Remove 'invoke' tests and fix other tests by mounting the 'basic'
middleware when required.

Update the documentation and examples to make use of the 'basic'
middleware.