vsgi: Permit multiple 'listen' calls (fix #132)
By calling multiple times 'listen', it's possible for a given
implementation to listen from multiple sources.

For SCGI:

 - rename and expose a 'socket-service' property
 - initialize 'SocketService' in a 'construct' block
 - only set 'backlog' if specified instead of overwritting an existing
   value

For HTTP:

 - initialize the server once
 - check if initialization specific options are provided and throw
   appropriate errors if the server is already initialized
 - set the 'tls-certificate' for 'https' whether it is specified or not
   because 'https' can be specified further
 - remove 'SERVER_SERVER_HEADER', we don't need that varidic hack
   anymore

For CGI and FastCGI:

Throw an error if the server is already listening on standard input: the
CGI implementation can only be initialized once.

Add a test to ensure that multiple calls to 'FastCGI.Server.listen'
fail.