Add a 'safely' middleware
The main purpose of this middleware is to provide a way to warn for
non-status errors instead of handling the upstream.

    app.get ("/", safely ((req, res, next, ctx) => {
        return res.expand_utf8 ("Hello world!"); // warn for 'GLib.IOError'
    }));

When dealing with more complex applications, one is likely to want to
handle error conditions with more granularity, and for that to happen,
it's great to let 'valac' warn.