negotiation: Check for compound subtypes in 'accept'
For subtypes, all that matter is that the accepted subtypes form a
subset of the proposed ones. For example, if the server is proposing
'application/vnd.api+json', it has to consider 'application/json'
acceptable.

Instead of checking specific cases (e.g. '*/*' and '/*'), we do the
following:

 1. check for type match with wildcard '*' or return 'false'
 2. check for subtypes with wildcard '*', all accepted subtypes must be
    satisfied by a proposed one or return 'false'
 3. return 'true' otherwise

Add a test to cover compound subtypes.