sse: Write a comment in the stream every 15 seconds
W3C recommends to keep the connection alive by sending a comment over
the event stream since proxies are likely to close inactive connections.

To prevent keeping the connection alive for sending non-events, we check
the refcount of the response and remove the source if we are the only
owner. In practice, it means that the connection will be open for at
least 15 seconds, but it's okay since event stream are typically used
for much longer.