vsgi: Provide a pipe between the master and worker
The master produce, the worker consume, everyone's happy.

Replace the list of 'GLib.Pid' by a list of 'Worker' objects that bundle
both a process identifier and a writable pipe. The server itself hold
the readable part of the pipe, which is 'null' on the master.

Use 'Posix.FD_CLOEXEC' to prevent pipes from leaking in other forks.

The communication protocol is left to the end-user, although sending
'GLib.Variant' messages would be recommended.

Add a basic test to ensure that the pipe works well.