vsgi: Provide 'stop' with really graceful shutdown
Provide a new abstract method in 'Server' to stop accepting incoming
connections. Using the 'shutdown' signal was not appropriate because it
executed once the main loop terminated and no new connections would be
accepted anyway (assuming 'Socket.accept_async').

Release before calling 'stop' as we hold after a successful 'start'.

Hold and release for each connection so that the server is kept alive as
long as necessary. This is basically keeping the server alive even if it
has itself released.

Use a low priority for handling UNIX signals such that stopping and
releasing are the really last operation performed during an interation.

Call 'dispatch' in a idle function for the CGI implementation such that
the 'SIGTERM' signal is already covered when it start executing.